Main functions

Motion correction

RetrospectiveMotionCorrectionMRI.motion_corrected_reconstruction — Method

motion_corrected_reconstruction(F, d, u, θ, options)

Performs retrospective motion correction of some data d. F is the Fourier operator, initialized via the package UtilitiesForMRI. The initial estimates for image and motion parameters are u and θ. The minimization options are passed via the routine motion_correction_options.

See this section for more details on the solution algorithm.

RetrospectiveMotionCorrectionMRI.motion_correction_options — Method

motion_correction_options(; image_reconstruction_options,
                            parameter_estimation_options,
                            niter,
                            verbose, fun_history)

Return motion correction options for alternating rigid motion estimation and image reconstruction:

image_reconstruction_options: set via the corresponding option routine

parameter_estimation_options: set via the corresponding option routine

niter: number of outer two-step loops

verbose, fun_history: for debugging purposes

Note: for more details consult this section.

Image reconstruction

RetrospectiveMotionCorrectionMRI.image_reconstruction — Method

image_reconstruction(F, d, initial_estimate, options)

Performs image reconstruction by fitting the data d through a linear operator F (e.g. the Fourier transform). initial_estimate sets the initial guess. The minimization options are passed via the routine image_reconstruction_options.

See this section for more details on the solution algorithm.

RetrospectiveMotionCorrectionMRI.image_reconstruction_options — Method

image_reconstruction_options(; prox::AbstractProximableFunction,
                               Lipschitz_constant=nothing,
                               niter_estimate_Lipschitz=nothing
                               Nesterov=true,
                               reset_counter=nothing,
                               niter=nothing,
                               verbose=false,
                               fun_history=false)

Returns image reconstruction options for the routine image_reconstruction:

prox: regularization function (for which a proximal operator is implemented)

Lipschitz_constant: Lipschitz constant of a smooth objective

niter_estimate_Lipschitz: when set to an Integer, the iterative power method is invoked in order to estimate the Lipschitz constant with the specified number of iteration

Nesterov: allows Nesterov acceleration (default)

reset_counter: number of iterations after which the Nesterov acceleration is reset

niter: number of iterations

verbose, fun_history: for debugging purposes

Note: for more details on each of these parameters, consult this section.

Rigid motion parameter estimation

RetrospectiveMotionCorrectionMRI.parameter_estimation — Method

parameter_estimation(F, u, d, initial_estimate::AbstractArray{T}, options)

Solves the rigid-motion parameter estimation optimization problem described here. F is a Fourier linear operator, initialized with the package UtilitiesForMRI (see Section Getting started for some examples on how to do it). u is a fixed known image, d given data, and initial_estimate a starting guess for the rigid motion parameters.

For optimization options, refer to parameter_estimation_options.

RetrospectiveMotionCorrectionMRI.parameter_estimation_options — Method

parameter_estimation_options(; niter=10,
                               steplength=1f0,
                               λ=0f0,
                               scaling_diagonal=0f0, scaling_mean=0f0, scaling_id=0f0,
                               reg_matrix=nothing,
                               interp_matrix=nothing,
                               verbose=false,
                               fun_history=false)

Returns parameter estimation options for the optimization problem underlying the solver parameter_estimation:

niter: number of iterations

steplength

λ: weight of regularization

scaling_diagonal, scaling_mean, scaling_id: parameters associated to the pseudo-Hessian of the objective

reg_matrix: regularization weight matrix

interp_matrix: interpolation matrix

verbose, fun_history: for debugging purposes

Note: for more details on each of these parameters, consult this section.

Rigid registration utilities

RetrospectiveMotionCorrectionMRI.rigid_registration — Method

rigid_registration(u_moving, u_fixed, θ, options;
                   spatial_geometry=nothing,
                   nscales=1)

Performs the rigid registration of the 3D image u_moving with the fixed 3D image u_fixed. The user can input a rigid motion parameter θ (nothing in case no prior knowledge is available), and options via rigid_registration_options. In case the 3D image are associated to a specific spatial discretization, it can be specified with the keyword argument spatial_geometry. Multiscale acceleration can be specified with by setting the number of levels with the keyword nscales (in this case, each scaled subproblem solver runs options.niter iterations).

RetrospectiveMotionCorrectionMRI.rigid_registration_options — Method

rigid_registration_options(; niter=10, verbose=false, fun_history=false)

Returns options for the rigid registration routine. niter sets the number of iterations, verbose=true can be used for debugging purposes, and fun_history=true allows the storage of the objective values at each iteration.